Clean-up Profile completion modules and undue CommunityPost refactoring
