--- src/containers/community/CommunityPosts/index.js
@@ -1,2 +1,0 @@
-import connector from './connector'
-export default connector
